Split system specs into individual files to make it easier to maintain tests
