+++ spec/system/edit_sign_feature_spec.rb
@@ -3,7 +3,6 @@ require "rails_helper"
 RSpec.describe "Editing a sign", type: :system do
   include FileUploads
   include WaitForAjax
-  include ActionView::Helpers::NumberHelper
 
   let!(:topic) { FactoryBot.create(:topic) }
   let(:user) { FactoryBot.create(:user, :approved) }
@@ -161,199 +160,6 @@ RSpec.describe "Editing a sign", type: :system do
       it {
         expect(subject).to have_selector("source[src*='/videos']", count: 3, visible: false)
       }
-    end
-  end
-
-  it "shows usage examples for a sign" do
-    attachment_1 = fixture_file_upload("spec/fixtures/small.mp4")
-    attachment_2 = fixture_file_upload("spec/fixtures/medium.mp4")
-    sign.usage_examples.attach(attachment_1)
-    sign.usage_examples.attach(attachment_2)
-    visit edit_sign_path(sign)
-
-    within ".sign-usage-examples" do
-      expect(page).to have_content attachment_1.original_filename
-      expect(page).to have_content number_to_human_size(attachment_1.size)
-      expect(page).to have_content attachment_2.original_filename
-      expect(page).to have_content number_to_human_size(attachment_2.size)
-      expect(page).to have_button "Remove File", count: 2
-    end
-  end
-
-  it "can update the description of a usage example", uses_javascript: true do
-    attachment = fixture_file_upload("spec/fixtures/small.mp4")
-    sign.usage_examples.attach(attachment)
-    visit edit_sign_path(sign)
-    desc = "New usage example description"
-
-    within(".sign-usage-examples li") do
-      field = find_field(:description)
-      field.send_keys desc, :return
-    end
-
-    wait_for_ajax
-    visit current_path
-
-    within(".sign-usage-examples li") do
-      expect(page).to have_field :description, with: desc
-    end
-  end
-
-  it "can update the description of a usage example without JS" do
-    attachment = fixture_file_upload("spec/fixtures/small.mp4")
-    sign.usage_examples.attach(attachment)
-    visit edit_sign_path(sign)
-    desc = "New usage example description"
-
-    within(".sign-usage-examples li") do
-      fill_in :description, with: desc
-      page.find("input[type=submit]", visible: false).click
-    end
-
-    visit current_path
-
-    within(".sign-usage-examples li") do
-      expect(page).to have_field :description, with: desc
-    end
-  end
-
-  it "can remove a usage example", uses_javascript: true do
-    attachment = fixture_file_upload("spec/fixtures/small.mp4")
-    sign.usage_examples.attach(attachment)
-    visit edit_sign_path(sign)
-
-    expect do
-      within(".sign-usage-examples li") { click_on("Remove File") }
-      expect(page).not_to have_selector(".sign-usage-examples li")
-    end.to change(sign.usage_examples, :count).by(-1)
-
-    expect(page).not_to have_selector ".sign-usage-examples li"
-  end
-
-  it "can remove a usage example without JS", uses_javascript: false do
-    attachment = fixture_file_upload("spec/fixtures/small.mp4")
-    sign.usage_examples.attach(attachment)
-    visit edit_sign_path(sign)
-
-    expect do
-      within(".sign-usage-examples li") { find(".show-if-no-js", text: "Remove File").click }
-    end.to change(sign.usage_examples, :count).by(-1)
-
-    expect(page).not_to have_selector ".sign-usage-examples li"
-  end
-
-  it "can specify usage examples for a sign", uses_javascript: true, upload_mode: :uppy do
-    within ".usage-examples" do
-      click_on "Upload files"
-
-      # Invalid file has an error
-      choose_file Rails.root.join("spec/fixtures/dummy.exe"), selector: "files[]", visible: false, match: :first
-      expect(page).to have_content "You can only upload: video/*"
-
-      # Block too many files
-      choose_file selector: "files[]", visible: false, match: :first
-      click_on "Add more"
-      choose_file Rails.root.join("spec/fixtures/small.mp4"), selector: "files[]", visible: false, match: :first
-      expect(page).not_to have_button "Add more"
-
-      click_on "Upload 2 files"
-    end
-
-    expect(page).to have_selector(".sign-usage-examples > *", count: 2)
-  end
-
-  it "can drag and drop usage examples for a sign", uses_javascript: true, upload_mode: :uppy do
-    within ".usage-examples" do
-      click_on "Upload files"
-      expect(page).to have_selector(".uppy-Dashboard")
-      drop_file Rails.root.join("spec/fixtures/dummy.exe"), selector: ".uppy-Dashboard"
-      expect(page).to have_content "You can only upload: video/*"
-
-      drop_file(selector: ".uppy-Dashboard")
-      drop_file("spec/fixtures/small.mp4", selector: ".uppy-Dashboard")
-      drop_file("spec/fixtures/medium.mp4", selector: ".uppy-Dashboard")
-      expect(page).to have_content "You can only upload 2 files"
-
-      click_on "Upload 2 files"
-    end
-
-    expect(page).to have_selector(".sign-usage-examples > *", count: 2)
-  end
-
-  it "can specify usage examples for a sign without javascript", uses_javacript: false do
-    within(".usage-examples") { choose_file }
-    click_on "Update Sign"
-    expect(page).to have_content I18n.t("signs.update.success")
-    visit edit_sign_path(sign)
-    expect(page).to have_selector(".sign-usage-examples > *")
-  end
-
-  it "can specify usage examples for a sign using the legacy uploader", uses_javascript: true, upload_mode: :legacy do
-    within ".usage-examples" do
-      choose_file
-      expect(page).to have_selector(".sign-usage-examples > *", count: 1)
-      choose_file Rails.root.join("spec/fixtures/small.mp4")
-      expect(page).to have_selector(".sign-usage-examples > *", count: 2)
-    end
-  end
-
-  it "can specify illustrations for a sign", uses_javascript: true, upload_mode: :uppy do
-    within ".illustrations" do
-      click_on "Upload files"
-
-      # Invalid file has an error
-      choose_file Rails.root.join("spec/fixtures/dummy.exe"), selector: "files[]", visible: false, match: :first
-      expect(page).to have_content "You can only upload: image/*"
-
-      # Block too many files
-      choose_file Rails.root.join("spec/fixtures/image.png"), selector: "files[]", visible: false, match: :first
-      click_on "Add more"
-      choose_file Rails.root.join("spec/fixtures/image.jpeg"), selector: "files[]", visible: false, match: :first
-      click_on "Add more"
-      choose_file Rails.root.join("spec/fixtures/image.gif"), selector: "files[]", visible: false, match: :first
-      expect(page).not_to have_button "Add more"
-
-      click_on "Upload 3 files"
-    end
-
-    expect(page).to have_selector(".sign-illustrations > *", count: 3)
-  end
-
-  it "can drag and drop illustrations for a sign", uses_javascript: true, upload_mode: :uppy do
-    within ".illustrations" do
-      click_on "Upload files"
-      expect(page).to have_selector(".uppy-Dashboard")
-      drop_file Rails.root.join("spec/fixtures/dummy.exe"), selector: ".uppy-Dashboard"
-      expect(page).to have_content "You can only upload: image/*"
-
-      drop_file("spec/fixtures/image.png", selector: ".uppy-Dashboard")
-      drop_file("spec/fixtures/image.jpeg", selector: ".uppy-Dashboard")
-      drop_file("spec/fixtures/image.gif", selector: ".uppy-Dashboard")
-      drop_file("spec/fixtures/image.webp", selector: ".uppy-Dashboard")
-      expect(page).to have_content "You can only upload 3 files"
-
-      click_on "Upload 3 files"
-    end
-
-    expect(page).to have_selector(".sign-illustrations > *", count: 3)
-  end
-
-  it "can specify illustrations for a sign without javascript", uses_javacript: false do
-    within(".illustrations") { choose_file Rails.root.join("spec/fixtures/image.png") }
-    click_on "Update Sign"
-    expect(page).to have_content I18n.t("signs.update.success")
-    visit edit_sign_path(sign)
-    expect(page).to have_selector(".sign-illustrations > *")
-  end
-
-  it "can specify illustrations for a sign using the legacy uploader", uses_javascript: true, upload_mode: :legacy do
-    within ".illustrations" do
-      choose_file Rails.root.join("spec/fixtures/image.png")
-      expect(page).to have_selector(".sign-illustrations > *", count: 1)
-      choose_file Rails.root.join("spec/fixtures/image.jpeg")
-      expect(page).to have_selector(".sign-illustrations > *", count: 2)
-      choose_file Rails.root.join("spec/fixtures/image.gif")
-      expect(page).to have_selector(".sign-illustrations > *", count: 3)
     end
   end
